Support different Int/Float sizes in generics
This adds support for and fixes compiling generic code involving
integers and floats with different sizes. Before these changes the
compiler incorrectly assumed all integers and floats require 8 bytes of
stack space, resulting in undefined and broken behaviour when passing
e.g. an Int32 to a generic type parameter.

C structures are not yet supported and the compiler now emits an error
when passing such types to generic type parameters. Supporting such
types requires a more complicated setup, as we'd have to support
generating shapes for arbitrary types. While this is something I'd like
to support (also as a potential future optimization for regular types),
this is something we'll implement separately in the future.

As part of this the compiler also includes a change that restricts field
types of external classes to C types or value types, ensuring you can't
violate Inko's ownership constraints by sticking owned values or borrows
in an external class.

Changelog: added