Fix assigning types with ownership to placeholders
This problem was originally described in
https://github.com/inko-lang/inko/issues/610, and was supposed to be
fixed in commit c9c01acb. This however wasn't entirely the case, as
certain scenarios (e.g. using pointers) could still result in
placeholders being assigned the wrong types.

This commit fixes this, ensuring that comparing e.g. `Pointer[A]` with
`Pointer[B]` results in `B` being assigned `A` instead of `Pointer[A]`.
This in turn means `Array` no longer needs to use a `Pointer[UInt64]` to
work around this bug.

Changelog: fixed