Update Git Email mapping
For a while I used a dedicated alias hoping this would reduce Email
spam, but this isn't really the case. In addition, GitHub has a tendency
to use the primary Email address anyway when using web actions (e.g.
rebasing a pull request), so the regular Email address is out there
anyway.