Enable std.test colors for terminals only
With the introduction of Stdout.terminal?, we can now automatically
disable the use of colors for test output if the test suite isn't
connected to a terminal.

In addition, the use of colors can be forcefully disabled by setting the
NO_COLOR environment variable to a non-empty value.

Changelog: changed