Fix capturing of C structures
Capturing C structures would result in them being captured as-is, which
makes using them rather in closures rather useless. Instead, closures
capture them as a pointer. This makes working with C structures in
closures dangerous as one has to ensure the closure doesn't outlive the
data, but such is the case when dealing with any form of C code.

Changelog: fixed