Reduce cost of creating process stacks
When spawning a process, we use mmap(2) to set up the memory used for
the process' stack. As part of this, we write some data to the start of
the stack, such as a pointer to the process itself. This can trigger a
page fault when the stack is new, resulting in the kernel committing the
first page. This in turn can be rather slow: somewhere between 5 and 10
microseconds, depending on how lucky you are.

To work around this, we now reserve a number of stacks per thread upon
startup, and we (volatile) write a dummy value to the private data
section of the stack. This allows hiding of the cost for the reserved
stacks, though depending on the number of processes spawned and how long
they stick around one can still run into this cost.

Changelog: performance