Reintroduce futures using Future and Promise
This commit reintroduces futures, allowing messages to send their
results back to the sender of the message, without the need for a
channel. Futures are split into two user facing types: Future and
Promise. A Future is a proxy for a value to be computed in the future,
while a Promise is used for specifying what that value is. Both a Future
and Promise can only be used once (e.g. Future.get consumes its
receiver). This in turn allows for a simple and efficient
implementation, rather than the more complex implementation of the old
Channel type.

The Channel type is re-implemented using a process, Future and Promise
pairs, and the new std.deque.Deque type. This implementation is an
_unbounded_ channel, as making a bounded channel using processes is a
little more tricky and not necessarily beneficial. The new Channel type
is just a regular type and doesn't use atomic reference counting,
requiring explicit calls to `Channel.clone` paired with a `recover` to
allow moving the channel between processes.

Changelog: changed