Tweak handling inlining of recursive calls a bit
When calling DateTime.new and possibly other methods, it was possible
for Int.unchecked_div to not be inlined due to it being flagged as a
recursive method. This commits improves this a bit by _not_ flagging a
method as recursive if the strongly connected components contain a call
that never returns. So for the chain `A -> B -> C -> A`, if `C` never
returns we don't flag `A` as recursive. In the case of DateTime.new,
this ensures that Int.unchecked_div _is_ inlined.