Fix unreliable ordering when specializing types
When specializing types we'd order over the contents of a HashSet, but
Rust's HashMap and HashSet types don't enforce a specific order,
resulting in non-deterministic behaviour. This in turn would result in
incremental caches being flushed randomly.

To fix this, we now use IndexSet type from the indexmap crate, which
_does_ enforce a consistent order. In addition, our custom IndexMap type
is replaced with the IndexMap type from the same crate.

This fixes https://github.com/inko-lang/inko/issues/765.

Changelog: fixed