Fix removing unreachable blocks after inlining
This ensures that unreachable blocks produced after inlining (but also
just in general) are properly removed. Specifically, if there's a path
A -> B -> C and A is unreachable, we remove A and the edge from A to B,
then (re)schedule B. If B is now unreachable, we do the same work, then
for C, and so on.

This fixes https://github.com/inko-lang/inko/issues/767.