Implement socket operations in Inko
Instead of relying on Rust code exposed through runtime library
functions, all socket operations are now implemented in Inko's standard
library. For TLS sockets we rely on callbacks for performing TLS reads
and writes, such that we don't need to expose any socket logic to the
TLS code.

By making these changes we're now fully in control of the socket logic,
the functionality we want to support, and how we want to support it. In
addition, we can drop the dependency on the socket2 crate.

In case of TLS sockets this may also result in a more efficient
implementation, as encountering EAGAIN/EWOULDBLOCK only retries the
low-level read/write instead of the entire TLS operation, though it's
difficult to say whether this makes a measurable difference.

This fixes https://github.com/inko-lang/inko/issues/752.

Changelog: changed