Fix underflow when removing unreachable blocks
When removing unreachable basic blocks we maintain a shift map. This map
keeps track of how many times a block ID must be shifted to the left,
based on the removed blocks before it.

The implementation didn't take into account that it would potentially
visit the same block multiple times, resulting in incorrect shift
amounts.

This commit fixes this by simplifying the implementation, which comes
with the added benefit of being able to remove unreachable blocks that
are part of a loop (e.g. A -> B -> A), something the previous
implementation didn't handle.

This fixes https://github.com/inko-lang/inko/issues/774.

Changelog: fixed