Don't produce redundant unreachable warnings
When comments are included, we no longer lower these to anything in
sequences of values. When used as single expressions (e.g. a value of a
`let`), we lower them directly to Nil instead of Noop.

The result is that we no longer produce redundant "this code is
unreachable" diagnostics when a comment follows something that makes it
unreachable (e.g. a return).

This fixes https://github.com/inko-lang/inko/issues/772.

Changelog: fixed