Refactor std.json.ObjectParser
The number of methods it offers has been reduced and it no longer stores
a PullParser, instead it's passed in through ObjectParser.parse. In
addition, flagging rules as required/optional is changed as well.
PullParser.object is changed to take a closure that receives an
ObjectParser, which is then used to parse the input stream.

While the resulting API is still a bit allocation heavy, that's
something we'll have to address separately by e.g. stack allocating data
and inlining closures.

This fixes https://github.com/inko-lang/inko/issues/719.

Changelog: changed