Set locations for instructions that call functions
This ensures that we set the LLVM debug location for all instructions
that may perform a function call, including calls to malloc(). This
won't affect the stack traces generated by the standard library, but it
does ensure that profiling applications (e.g. using Linux' perf) operate
on the correct stack traces.

Changelog: changed