Make various types inline types
For types such as tuples, Option, Result, and various others, there's no
need to allocate them on the heap. In fact, for frequently used types
such as Option and Result this does nothing but harm performance. To
resolve this, various types are marked as `inline` such that they are
allocated on the stack.

Changelog: performance