Make various types inline types
For types such as tuples, Option, Result, and various others, there's no
need to allocate them on the heap. In fact, for frequently used types
such as Option and Result this does nothing but harm performance. To
resolve this, various types are marked as `inline` such that they are
allocated on the stack.

IO types such as ReadOnlyFile and Stdout remain heap types for the time
being, as this allows casting them to e.g. the Write trait. This in turn
is useful if some deeply nested type needs to act on different Write
types, but you don't want to litter generic type arguments all over the
place.

Changelog: performance