Add support for DateTime parsing and formatting
This adds the following methods to the DateTime type:

- DateTime.parse
- DateTime.format
- DateTime.to_iso8601
- DateTime.to_rfc2822

These methods support locale-aware parsing and formatting of DateTime
values, with the standard library providing support for English, Dutch
and Japanese (with the ability to add more in the future). Locale data
is located in locale-specific modules such as std.locale.en and
std.locale.ja, such that this data is only included if necessary.

The parsing and formatting is done according to a formatting string,
which uses a syntax based on that of strftime(2)/strptime(2)

This fixes https://github.com/inko-lang/inko/issues/677
and fixes https://github.com/inko-lang/inko/issues/678.

Changelog: added