Replace the `class` keyword with `type`
Now that types are no longer guaranteed to go on the heap, the use of
"class" can be confusing as this is typically used for types that go on
the heap. In addition, our types weren't really traditional classes to
begin with due to the lack of support for inheritance.

The `class` keyword is still supported but is now deprecated, and the
`fmt` command automatically replaces it with the new `type` keyword.

This fixes https://github.com/inko-lang/inko/issues/779.

Changelog: changed