Add std.net.dns for performing DNS queries
This adds the std.net.dns module providing the Resolver type for
performing DNS queries. Currently the only type of query supported is
resolving a hostname into a list of IP addresses, but this may be
expanded upon in the future.

For FreeBSD and macOS, the implementation uses getaddrinfo(). For Linux,
if systemd-resolve is installed its varlink API is used with a fallback
to getaddrinfo(). The benefit of using the varlink API is that it allows
us to make use of Inko's non-blocking sockets.

This fixes https://github.com/inko-lang/inko/issues/735.

Changelog: added