Remove the IpAddresses type
Using multiple IP addresses in a particular order is really only
relevant for TcpClient.new and TcpClient.with_timeout. Binding a socket
is pretty much always done with a single IP address, and when multiple
IPs _are_ used it's likely that you want to bind to _all_ of them, not
the first one that works. For the low-level Socket type we don't do
anything special either, as it's up to consumers of that type to decide
how they want to connect when handling multiple IPs.