Fix Channel dropping values in certain conditions
The type std.sync.Channel is built on top of the Future and Promise
types. The implementation of Promise.set was such that if it's
disconnected prior to a write, the value would be dropped silently.
Combined with how Channel was implemented, this could result in sent
values being discarded.

To resolve this, Promise.set now returns the value to set if its
corresponding Future is already dropped, allowing callers to act
accordingly. The implementation of Channel in turn is changed to take
advantage of this, and to ensure values are always resolved in FIFO
order.

Changelog: fixed