Always consume self in Future.get_until
Future.get_until now always consumes self and returns an Option[uni T]
instead of Result[uni T, Future[T]]. The reason for this is that the use
of a timeout makes this method very racy and it's difficult to
efficiently handle cases where an Error is returned. For example, if a
value is written to a Promise _just_ after Future.get_until returns, one
would have to resort to a busy loop to observe the value at some point.

Consuming self in Future.get_until means that the Future is dropped upon
a timeout, disconnecting its Promise and forcing calls to Promise.set to
either discard the value or retry the operation in some way (e.g. by
using a new Future/Promise pair).

For more details, refer to
https://github.com/inko-lang/inko/issues/795#issuecomment-2568612941
and the comments that follow it.

Changelog: changed