Fix use after free when dealing with timeouts
This refactors the timeouts handling code to ultimately resolve its
soundness issues. That is, it was possible for a process to be dropped
while the timeouts thread still had one or more entries in the timeouts
heap. This would then result in this thread potentially trying to
reschedule a process that no longer exists. The result would either be a
crash or a hang, depending on how lucky (or not) you are.

This fixes https://github.com/inko-lang/inko/issues/796.

Changelog: fixed