Enable LLVM optimizations
This enables a set of LLVM optimization passes for the "balanced" and
"aggressive" profiles. Both are based on the "default<O2>" pipeline,
with the removal of some irrelevant passes. In the case of the
"balanced" profile we also remove some additional passes that aren't
likely to be useful in most cases.

This fixes https://github.com/inko-lang/inko/issues/595.

Changelog: added