Replace Debian containers with Fedora
Building the runtime library using Fedora should be fine as the
resulting static library doesn't dictate a specific glibc version for
GNU targets.

In addition, while we don't want to use the ARM64 runners for tests at
this stage, we _can_ use them for the runtimes as this should simplify
things a bit and lets us remove the Debian containers.

This is of course all assuming it works, which is always a bit of a
gamble with GitHub Actions.