Add support for stack allocated unique types
A unique type is a type that is allocated on the stack and subject to
move semantics (i.e. it's not copied), but severely restricts borrows.
This in turn makes it safe to assign fields new values (unlike with
inline types), as the assignments are always correctly observed (because
they always act on the same reference). Unique types are useful when
creating mutable state machines and streams such as files, sockets,
and iterators.

Similar to borrows of `uni T` _values_, one can use borrows of unique
types as the receiver of a method. For example, if a unique type is
stored in a field it's exposed as a `ref T` or `mut T`, and one can
still call methods on such a borrow. What isn't possible is assigning
such borrows, passing them around as arguments or returning them.

Refer to https://github.com/inko-lang/inko/issues/776 for more
information.

Changelog: changed