Expose the inner type for BufferedReader/Writer
This changes the types BufferedReader and BufferedWriter such that the
field storing the inner Read/Write type is public. This is useful when
one wants to move some type into these types, while still being able to
interact with it (e.g. by setting a timeout in the case of sockets).

Changelog: changed