Apply compiler cleanups from the unique types work
This applies various cleanups/tweaks to the compiler code that were
pulled from the work on unique types/borrowing as part of
https://github.com/inko-lang/inko/issues/804 and
https://github.com/inko-lang/inko/pull/798.

A change worth mentioning is that we no longer allow default methods to
capture `self` borrows in closures. The reason for this is that if the
implementing type is an `inline` type, the closures would capture a copy
of the stack data. The user in turn may expect any mutations on a
captured self to propagate back to the original type, which in this case
won't be true.