Add support for mutable and immutable fields
Similar to local variables, fields now default to not allowing
assignments (= their binding is immutable) and require explicit use of
the `mut` keyword to allow assignments:

    type Example {
      let mut @number: Int
    }

This makes it more difficult to assign new values by accident,
especially for public fields exposed through a library. In addition, it
opens the door to optimizations for fields that can't be assigned new
values, such as referring to them through a pointer when they contain an
`inline` type (instead of copying).

Changelog: changed