Default to debug builds and remove --opt
We now (once again) default to debug builds instead of release builds,
as slow release builds during development are annoying. In addition, the
--opt flag is removed in favour of a --release flag as there was no
practical/meaningful difference between --opt=balanced and
--opt=aggressive.

I experimented with printing a warning when building a debug build, but
ultimately found this too annoying. In addition, the output being
located in ./build/debug should be enough of a hint that the output is a
debug (aka slow) build.

This fixes https://github.com/inko-lang/inko/issues/825.

Changelog: changed