Warn for unused method results
The compiler now produces a warning when calling a method for which the
return value is unused. Such warnings are added based on a heuristic
rather than relying on user-provided annotations, as I'm not a fan of an
annotation based approach (= they end up being littered everywhere). The
heuristic is simple and as follows:

- We only warn when returning _owned_ values
- We don't warn when returning an Option, _unless_ the receiver of the
  call is also Option (e.g. the result of Option.map)
- We don't warn when returning Nil or Bool, as these types are pretty
  much always safe to ignore (e.g. the result of Set.insert)

This fixes https://github.com/inko-lang/inko/issues/809.

Changelog: added