Allow unsendable results when they're unused
When calling a method on a `uni T` value, we now allow the call even if
the return value isn't sendable if and only if the result is guaranteed
to never be used by the caller. This removes the need for explicit
recovery in a variety of cases.

As part of this change the drop semantics are slightly adjusted: unused
return values and values assigned to the `_` variable are dropped
immediately instead of at the end of the scope. This is necessary to
ensure that the caller won't drop data potentially containing borrows to
a value that is moved to another process.

This fixes https://github.com/inko-lang/inko/issues/612.

Changelog: added