Disallow moving of receivers as arguments
For expressions such as `x.y(x)` where `y` is a method that _borrows_
its receiver (= `fn` or `fn mut`), we reject the call if the receiver is
also moved as part of an argument.

While this particular pattern is fine for heap and inline values, when
the receiver is a unique value it can result in concurrent access to
shared data. For example, if `x` is a `uni T` and `y` moves its argument
to another process, _and_ `y` were to somehow return data from within
`x`, then both the calling and the new process may end up acting on the
same data (i.e. the caller decrements some borrow counts while the new
process increments them).

Even if this were sound in all cases, it's such a weird thing to both
borrow and move a receiver that it should just be rejected for every
instance where the receiver is an owned value.

This fixes https://github.com/inko-lang/inko/issues/829.

Changelog: fixed