Don't allow inline types to subtype traits
This fixes a bug in the type checker that would incorrectly allow inline
types to subtype trait values. For example, code such as this was
allowed:

    type inline Stream[T] {
      ...
    }

    trait Foo {
      fn iter -> Iter[Int]
    }

    type inline Bar {}

    impl Foo for Bar {
      fn iter -> Stream[Int] {
        ...
      }
    }

This isn't correct because traits used as values are boxed and addressed
through a pointer, which isn't the case for inline types.

Changelog: fixed