Add support for for loops
This adds support for a `for` expression, used to make it easier to
iterate over the values yielded by an iterator. For example, instead of
this:

    [10, 20, 30].iter.each(fn (num) { out.print(num) })

You'd write this instead:

    for num in [10, 20, 30].iter { out.print(num) }

The binding (`num` in the above example) is in fact a pattern, so you
can also write things like this:

    for (key, value) in [('name', 'Alice')].iter {
      print(key)
      print(value)
    }

A `for` loop is just syntax sugar and compiles down to a `loop` combined
with a `match`. The value to iterate over is expected to define the
method `into_iter` which in turn is expected to return a type that
implements `std.iter.Iter`. We don't explicitly check for that so in
theory any type is fine provided it implements `next` such that it
returns an `Option[T]`.

A big benefit of `for` loops over closures is that they don't need to
capture data, making it easier to work with e.g. `uni T` values as part
of a loop. In addition, you can break/next the `for` loop while with the
old `Iter.each` you could only skip to the next iteration (using
`return`), but not abort the iteration entirely.

As part of this change, the following methods are removed:

- std.iter.Iter.each
- std.iter.Iter.each_with_index
- std.iter.Iter.try_each

These methods are removed as they are redundant now that we have `for`
loops.

This fixes https://github.com/inko-lang/inko/issues/817.

Changelog: added