Require LLVM 18 or newer
In addition we now handle newer LLVM versions by falling back to
default<O2> (along with logging a warning message) if the list of passes
isn't supported by whatever newer version is used.

With this fallback in place we should also be able to support LLVM 19
and newer, though for the time being we'll still only officially support
LLVM 18 as that's the version we test against.

This fixes https://github.com/inko-lang/inko/issues/816.

Changelog: changed