Support verifying of MIR
This changes the --verify-llvm flag into a generic --verify flag that
also enables verifications of other parts of the compilation output. For
example, we now check for duplicate modules and methods if this flag is
specified. This in turn revealed a bug where InlineRef and InlineMut
shapes were coalesced into the same type in certain cases, resulting in
duplicate methods being present.

Changelog: fixed