Fix inferring closures based on how they're called
We now correctly infer the arguments of a closure according to any
"call" calls. The cause was that comparing placeholder A with
placeholder B would only assign A to B, but not B to A. This then meant
that assignments to A (through inference based on "call" calls) would
not propagate to placeholder B, resulting in the type specializer
producing incorrectly specialized methods.

This fixes https://github.com/inko-lang/inko/issues/822.

Changelog: fixed