Implement ByteArray in Inko
The implementation of ByteArray is now done entirely in Inko similar to
the Array type, instead of ByteArray being a type provided by the Rust
runtime library.

For users this won't matter much, but it makes maintaining and extending
the code easier as changes to ByteArray won't require a combination of
compiler, standard library _and_ runtime library changes.

Changelog: changed