Implement String in Inko
The String type is now implemented in Inko itself instead of being
defined by the runtime library. We still use a few Rust functions for
some string logic, such as converting a string to upper/lower case, but
everything else is done in Inko. To pass string-like data between the
standard and runtime library, the runtime library defines a
PrimitiveString type which is essentially a (pointer, size) pair with
contextual ownership.

As part of this change, string literals that use interpolation are
compiled a little different. Prior to this commit, such strings made use
of the inko_string_concat() runtime library function. Instead such
literals now use std.string.StringBuffer such that we can centralize the
logic in a single place. This means that this:

    let x = 'foo${10}bar'

Is compiled into (roughly) the following:

    let x = {
      let buf = StringBuffer.with_capacity(3)

      buf.push('foo')
      buf.push(10.to_string)
      buf.push('bar')
      buf.to_string
    }

As part of this work std.env is also changed a bit to make less use of
Rust functions and instead goes through libc directly for several of its
methods.

Changelog: changed