Refactor slicing API to not require copying
Slicing a String or ByteArray no longer involves creating a copy of the
sliced range. Instead, the type std.bytes.Slice is returned. This type
is a "view" into any type that implements the "Bytes" trait. Slice is
stack allocated and doesn't copy any data or takes ownership of it,
instead it's a reference to existing data along with a range.

In addition, instead of slices being created using a start index and
size, a start and end index is used instead. This results in a less
clunky slicing API and makes it consistent with what most other
languages do.

Changelog: changed