Refactor how std.io.Write must be implemented
Instead of requiring "write_string" and "write_bytes" to be implemented,
the trait std.io.Write now requires a "write" method with the following
signature:

    fn pub mut write[B: Bytes](bytes: ref B) -> Result[Nil, Error]

This allows writing of any type that implements the Bytes trait, such as
String, ByteArray and Slice. This makes it possible to write not only a
buffer as a whole, but also a sub slice (e.g. the first N bytes) using
the new slicing API. It also results in an easier to use API, both when
implementing the type and when using it. For example, instead of this:

    file.write_bytes(bytes)
    file.write_string('...')

You write the following:

    file.write(bytes)
    file.write('...')

Changelog: changed