Revert batched hashing of String/ByteArray
In 03b129b5 the implementation of String.hash and ByteArray.hash was
changed to hash bytes in chunks of 2/4/8 bytes at a time. While this
speeds up the hashing process, it also degrades the quality of the
resulting hash. This in turn can result in a significant increase in
hash collisions.

Refer to
https://github.com/inko-lang/inko/issues/836#issuecomment-2727514833 for
more details.

This reverts commit 03b129b58f303edc3f34eac6b4e316f35c1ea953.