Relax passing ref arguments to unique receivers
When calling a mutable method on a unique receiver, we now allow passing
immutable borrows that normally aren't sendable _if and only if_ the
receiver can't ever store a borrow.

A simple example of this is writing a bunch of bytes to a file: the
`write` method is defined as taking a `ref Bytes` but because it's a
mutating method we'd disallow this if the receiver is unique. The
changes in this commit means that such cases are now allowed.

This fixes https://github.com/inko-lang/inko/issues/841.

Changelog: added