Treat mut T arguments as sendable in certain cases
Given the expression `rec.method(arg)`, if `rec` never stores any
borrows and `arg` is a mutable borrow of a type that only stores value
types, we allow passing the argument if `rec` is a unique value. This is
safe because `arg` can't be aliased by `rec` because it doesn't store
any borrows, and `arg` can't alias `rec` because the values in `arg` are
all value types.

An example of where this is useful is the following:

    socket.read(into: bytes, size: 32)

If `socket` is a `uni TcpClient` then this expression would previously
be rejected. With the changes of this commit this is now allowed
because:

1. A TcpClient doesn't borrow anything, meaning it can't alias `bytes`
2. A ByteArray contains only value types, meaning it can't alias the
   `TcpClient`

Changelog: added