Support waiting for multiple signals
This adds the type std.signal.Signals which supports adding and waiting
for multiple signals, without the need for manually spawning processes
to handle each signal.

This commit also adds some basic tests for the std.signal module, which
I apparently forgot to do when initially adding the module. We don't
have any tests for actually sending signals at this point, as I'm not
sure yet how we'd best expose that in a way not just exclusive to the
test suite.

Changelog: added