Change how executables are built
Instead of treating src/main.inko as the starting point of an
executable, the `inko build` command now treats _every_ file directly in
the src/ directory as the start/source of an executable, deriving the
output name from the source file. For example, src/foo.inko results in
the executable build/debug/foo. This makes it possible for projects to
produce multiple executables without resorting to a custom build
process.

The visibility rules for private symbols are changed to accommodate this
new setup better: files placed directly in src/ have access to symbols
defined in files in sub directories (e.g. src/foo/bar.inko), even if the
executable source file uses a name that doesn't match the root namespace
of the sub file.

As part of this change the --output option is removed from the `inko
build` command. This option only makes sense when building a single
executable, and keeping it around adds additional complexity to the
compiler that just isn't justified.

This fixes https://github.com/inko-lang/inko/issues/631 and fixes
https://github.com/inko-lang/inko/issues/625.

Changelog: changed