Add the `inko init` command
This command is used for setting up a basic Inko project, making it a
little easier to get started with Inko.

As part of this change the documentation is updated in various places to
use this new command and favor the use of `inko build` instead of `inko
run`.

The Dockerfile is also updated to install tar and Git, such that the
GitHub Actions workflow file (optionally) generated by `inko init`
doesn't need to include extra steps to install these dependencies that
are required by GitHub.

Changelog: added