Allow non-exact reads in BufferedReader
This adds the field/flag BufferedReader.exact which defaults to true.
When set to false, BufferedReader uses Read.read instead of
Read.read_exact, making it suitable for use with stream types such as
TcpClient. The trade-off is that the buffer may contain fewer bytes than
the desired capacity, and that thus more read() calls may be required
over time compared to when exact mode is enabled.

Changelog: added