Change Int overflow behavior for release builds
Instead of always performing checked arithmetic, Int's operator methods
such as Int.+ and Int.>> now change their behavior based on the mode
used. In debug mode, these methods use checked arithmetic while in
release mode they use two's complement wrapping, similar to Rust.

The reason for this change is that while checked arithmetic helps detect
errors, it comes at a non-trivial cost. The exact cost can vary from a
few percent to a lot more, depending on the chain of optimizations that
is inhibited by the presence of checked arithmetic. For a single
function that might not matter much, but for larger applications this
can result in a cascade of poorly optimized code, resulting in a
potentially significant overall reduction of performance.

This fixes https://github.com/inko-lang/inko/issues/855.

Changelog: performance