Remove the Contains trait
Unlike equality, the meaning and implementation of "does X contain Y"
may change based on the type of X. For some types Y is an Int, for other
types it's an entire String. Generalizing over this with a trait isn't
helpful, so let's just let types define `contains?` directly without the
need for a trait.

Changelog: changed