Tweak inlining rules for $dropper methods
Instead of _never_ inlining calls to $dropper, we now do the following:
if the type is a stack allocated type, _always_ inline, otherwise leave
it up to the inliner's heuristics.

By always inlining droppers for stack types we can avoid redundant drop
logic, such as when a stack type stores only copy values or is an enum
of which the constructors doesn't store any data (e.g. an Option.None).

Changelog: performance